pyproject.toml
[project]
name = "techflow-chat"
version = "0.1.0"
description = "TechFlow Electronics customer support chat system - multi-agent retention and routing"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.3.0",
    "langgraph>=0.2.0",
    "langchain-google-genai>=2.0.0",
    "langchain-community>=0.3.0",
    "faiss-cpu>=1.9.0",
    "sentence-transformers>=3.0.0",
    "pydantic>=2.0",
    "python-dotenv>=1.0.0",
    "streamlit>=1.28.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv]
dev-dependencies = ["pytest>=8.0", "pytest-asyncio>=0.24"]
